streaming-redpanda-postgres/requirements.txt
kafka-python>=2.0.2
psycopg[binary]>=3.1.18
pydantic>=2.0.0
python-dotenv>=1.0.0

typer>=0.12.0
rich>=13.0.0

pytest>=8.0.0
ruff>=0.5.0
